requirements.txt
psycopg2-binary
celery==5.2.6
Django==4.1.7
djangorestframework==3.13.1
django-redis==5.2.0
drf-spectacular==0.22.1
djangorestframework-simplejwt==5.0.0
flower==1.0.0
gunicorn==20.1.0
python-dotenv==0.19.2
django-mptt==0.13.4
whitenoise==6.2.0